AlignGraph/requirements.txt
biopython==1.79
click==8.1.3
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyQt6==6.3.1
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
